Resolve inline components before root component
During instantiation of a component, we might need access to the
meta-objects of its referenced objects in the QQmlPropertyCacheCreator.
In the case of inline components, those might have however not been
complete: If they contained alias properties (which change the shape of
the metaobject), we would not be aware of it, as those are only resolved
when QQmlComponentAndAliasResolver has run.
QQmlComponentAndAliasResolver runs however necessarily after the
QQmlPropertyCacheCreator, as the alias resolver needs the metaobject to
know about the non-alias properties.

We fix this by restructuring the passes, so that we first run both
passes for each inline component, and only afterwards for the component
containing them. This is done by making QQmlPropertyCacheCreator
resumable: resumableBuildMetaObjects processes one (inline) component at
a time, and remembers enough state to continue.

Fixes: QTBUG-91143
Fixes: QTBUG-85980
Change-Id: Ib0d0bc5ff6cf067b2443b2e0cfb5400a9bd96754
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>