Fix incorrect depth test state with QSGRenderNode::DepthAwareRendering
Not applicable to Qt 6. But in Qt 5 GL_DEPTH_TEST must be re-enabled
after the QSGRenderNode has done with its work, and the rendernode
specified DepthAwareRendering. For rendernodes without this flag
everything is fine, but when the flag is set, the usage of the
depth buffer, and so depth testing in the alpha pass, stays enabled,
which means leaving GL_DEPTH_TEST as disabled will lead to incorrect
rendering with semi-transparent nodes that come after the rendernode.

Fixes: QTBUG-92036
Change-Id: Ide9e862e3504ded31eba5ee2a4caf804b69812e3
Reviewed-by: Andy Nichols <andy.nichols@qt.io>