Selection support: support setting a QItemSelectionModel on TableView
Add support for assigning a QItemSelectionModel to TableView.
By doing so, delegate items that has a "required property selected"
defined will get this updated according to the state of the
selection model.

It's essential that the property is defined as "required". If not, the
property will simply be ignored by TableView. This is done to ensure
that existing applications that already has a "selected" property
defined, will continue to work as before, unaffected by the new
selection API.

[ChangeLog][QtQuick] TableView now supports selections by using an
ItemSelectionModel.

Task-number: QTBUG-74750
Change-Id: I4f4d75e9e65563b9aab0c54f3aa4aad2f6883952
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>