Native style: fix statically built apps that links to both widgets and controls
The c++ symbols are already namespaced into QQC2. But since there is no
namespace concept in obj-c, we use QT_MANGLE_NAMESPACE to wrap those
symbols into a "fake" namespace by concatinating symbol name and
namespace name together.

The problem is that QT_MANGLE_NAMESPACE only take the QT_NAMESPACE
into account, and not the additional QQC2 namespace. The result is
that the application will fail linking because of duplicate obj-c
symbols when you combine widgets and controls in a statically linked
application.

This patch will redefine the affected macros locally in
qquickmacstyle_mac, so that both the Qt namespace and the controls
namespace are taken into account.

Pick-to: 6.0 6.1 6.1.0
Fixes: QTBUG-92883
Change-Id: Ic2f84ad9c0005fdd2553729d897cee113a4ff24d
Reviewed-by: Morten Johan Sørvig <morten.sorvig@qt.io>