Optimize QQmlBinding::slowWrite()
That's not uncommon at all, and we don't need to convert the value back
and forth for it. If we have a metatype/void* representation already, we
can just create a QVariant from that and forward it to
writeValueProperty(). Even the QVariant is unnecessary, but that will be
resolved in a second step.

Change-Id: Ideae979a73276c9fba1bc19c10d799b8fe59dde5
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>