QQuickDeliveryAgent::event(): handle tablet events explicitly
QQuickWindow::event() forwards all input events to the DA. If it's not
handled there, it eventually gets handled at the bottom by forwarding to
QWindow::event() which then dispatches to QQuickWindow::tabletEvent()
and from there to da->deliverPointerEvent(); however the switch in
QWindow::event() is not meant to be the normal entry point for input
events anymore.  In a subsequent patch we will move the failsafe
"never allow any kind of grab to persist after release" into
QQuickWindow::event(), before that is reached; that in turn breaks
tst_PointHandler::tabletStylus(), showing that the point is ungrabbed on
release before PointHandler handles it and has a chance to emit
pointChanged.  So we must handle tablet events earlier, alongside
mouse and touch events.

Amends 68c103225f4e8bd6c1b18ef547108fd60f398c0f.

Pick-to: 6.1
Change-Id: Id7508e4bbbdbfe88d002d298c15ec727040f02ef
Reviewed-by: Shawn Rutledge <shawn.rutledge@qt.io>