Make tst_qquickflickable check scaling of drag threshold
We continue to enforce the drag threshold in local (transformed)
coordinates: if Flickable should wait until the user drags 12 pixels,
but it's scaled to half-size, it will start dragging after only 6 pixels
of movement, for example. But if it's also rotated, then the required
distance becomes a projection of the actual drag vector onto the vector
along which the Flickable should move past the (scaled) drag threshold.
I.e. if we rotate the Flickable's parent by 45°, the required distance
should be 12 / 2 * √2 ≈ 8.5; 9 pixels will exceed the threshold, then.
tst_qquickflickable::clickAndDragWhenTransformed is modified to prove it
(at least when a 2D scale is applied).  The test should be OK on hidpi
because qApp->styleHints()->startDragDistance() is already adjusted.

Change-Id: I035a462c75e7c5aba876a67c1b996bccab2c5364
Reviewed-by: Shawn Rutledge <shawn.rutledge@qt.io>