Change the way how QmlDevTools propagate include directories
QmlDevToolsPrivate links QmlDevTools using the INTERFACE linking type.
This means that all PUBLIC/INTERFACE include directories of the
QmlDevTools will be propagated to the targets that link
QmlDevToolsPrivate. Since QmlDevTools is defined as an INTERNAL_MODULE
it should not be linked directly, but only as an interface dependency
of the QmlDevToolsPrivate target. According to the statements above assume
that:
     target_include_directories(QmlDevTools PRIVATE ...)
     target_include_directories(QmlDevToolsPrivate INTERFACE ...)

could be squashed into:
     target_include_directories(QmlDevTools PUBLIC ...)

This is precondition for QTBUG-87775

Change-Id: Icd772f0900191f3972fa7432cdcb06838981b23c
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>