ShaderEffect docs: Clarify vertex output/input naming requirements
To be complemented by an rhi change that adds verification for this
and prints a warning with mismatching names. The common use case with
ShaderEffect is to provide a fragment shader only, while using the
default, built-in vertex shader. In this case one must be aware of
how to correctly declare the texture coordinates input, and this was
not very clear in the docs.

In short, we need to advise that the fragment shader input must be
declared as qt_TexCoord0 at location 0 in order to be portable. Just
having a matching location number but a different name is not
sufficient. (because with old GLSL versions the location qualifier
will not be there at all and matching is name based)

Pick-to: 6.1
Task-number: QTBUG-92500
Task-number: QTBUG-93370
Change-Id: I975557e9a754db8e7c69b86531a2d873f059c7ce
Reviewed-by: Christian Strømme <christian.stromme@qt.io>