QQuickDial: Keep value integer if everything is integer
If a user uses integer values for stepSize, from and to in a Dial, they
most likely want the actual values of the Dial to be integers, too.
Detect this condition, and store it in a new boolean member. If the
condition is met, we round the value in QQuickDialPrivate::valueAt
(which, due to floating point math might not be an exact integer).

As a drive-by, reorder the boolean members to introduce no additional
space overhead.

Pick-to: 6.1 5.15
Fixes: QTBUG-92214
Change-Id: If4633fae1d7d425ca7fb767c7284d6f8ea7ce78c
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>