macOS, Slider: draw a focus rect, and not a focus ring
In the mac style, we draw the focus ring for a slider
from QML around the handle. This has shown to be fragile, since
UIKit doesn't draw the handle in the center of the
rect we get when we query UIKit (using the
[cell knobRectFlipped:slider.isFlipped]). And then the
focus ring will be offset as well.

While we could try to adjust the rect manually from
QQuickMacStyle, this will only be quaranteed to work
on top of the current macOS SDK (macosx11.1). So for
that reason, this patch falls back to use the less
correct, but more future safe, solution of drawng the
focus ring as a rect around the handle instead.
While this doesn't look native, it still looks
better (and correct, for the uninformed) compared
to an eliptic ring that is misplaced on the target.

Fixes: QTBUG-93423
Pick-to: 6.1 6.0
Change-Id: Ibc00f12f9606cd66c8ff384967aa3ea0d7727964
Reviewed-by: Jan Arve Sæther <jan-arve.saether@qt.io>