ToolTip: use contentWidth of Text contentItem to account for newlines
By default, QQuickPopupItem uses the implicitWidth of its contentItem,
which is too large in the case of a ToolTip with newlines in the text.
In that case, contentWidth refers to the width of the text including
newlines, so we use that instead.

[ChangeLog][Controls][ToolTip] The implicit width of ToolTips now
accounts for newlines in the text. If you want to use the old behavior,
set ToolTip's contentWidth to implicitContentWidth.

Fixes: QTBUG-83630
Pick-to: 6.1 5.15
Change-Id: I7ca3805429acb68a13ead8f3545bb84a51fb1b72
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>