Extend qmlimportscanner to support response files on the command line
This is preparation for being able to list individual qml files on the
command line instead of requiring them to be in a .qrc file.
To avoid running into line length issues when there are a lot of files,
allow response files to be used to store arguments instead.
Use the common tool convention of a leading "@" before the name of the
response file.

Change-Id: Id91daf08a6022033a448e9ed98085df8b246a849
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>
Reviewed-by: Andrei Golubev <andrei.golubev@qt.io>