Implement accessibility for QQuickWidget
The accessibility tree for the Qt Quick content should
be rooted at the QQuickWidget, and not at the offscreen
QQuickWindow.

For this to be the case, several things must happen:
  - QQuickWindow must not report the child interfaces
  - QQuickWidget must report the child interfaces
  - The child interfaces must report the QQuickWidget as the parent

Create accessibility interfaces for QQuickWidget and
and QQuickWigetOffscreenWindow (which now gets a proper
subclass), where the QQuickWidget interface reports
the child interfaces and the QQuickWigetOffscreenWindow
reports no children

Change the code in QAccessibleQuickItem to use the
true (visible) window, where needed.

Fixes: QTBUG-67290
Change-Id: I387d0ef711138d248a8dd16eefc9839499b35eeb
Reviewed-by: Jan Arve Sæther <jan-arve.saether@qt.io>
Reviewed-by: Volker Hilsheimer <volker.hilsheimer@qt.io>