Actually test day of the week in test-case for day-of-week bug
The test-cases I added for QTBUG-78996 tested the root cause of the
bug, but neglected to test the symptom actually reported - the day of
the week was wrong. Add actual an day-of-week check to each (and
semicolons on the ends of statements).

This amends commit 290dc2da70655db9d4590600dfd37b825b81c1f8.

Task-number: QTBUG-78996
Change-Id: Ic7064ddd5a8e4fa0806a10d73cd5955c7be4a4bc
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>